androidApp: ScanTicketViewModel: Avoid scanning same QR multiple times
Untill an image proxy is closed, camerax avoids starting a new analysis. Use this
to our benefit and block the thread while we are fetching the attendee to verify the ticket.

Also add an intentional 2 seconds delay if we detected a QR code earlier to avoid scanning
the same QR again.

Signed-off-by: Aayush Gupta <aayushgupta219@gmail.com>